+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Include documentation in gem package
 * Automatically set a network for localhost
+* Add config option (proc) transform_result