+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Include documentation in gem package
 * Automatically set a network for localhost
 * Add config option (proc) transform_result
+* iso_code fallbacks to code and vice versa so you don't have to think about it (continent has code and rest has iso_code).
+  This only applies to method access, `[:code]` does not fallback.